restructured grid_2d_pixelization into its own structure package
