+++ test_autoarray/inversion/inversion/test_abstract.py
@@ -548,7 +548,7 @@ def test__errors_and_errors_with_covariance():
 def test__brightest_reconstruction_pixel_and_centre():
 
     mapper = aa.m.MockMapper(
-        source_mesh_grid=aa.Grid2DVoronoi.manual_slim(
+        source_mesh_grid=aa.Mesh2DVoronoi.manual_slim(
             [[1.0, 2.0], [3.0, 4.0], [5.0, 6.0], [5.0, 0.0]]
         )
     )
