+++ test_autoarray/inversion/mappers/test_abstract.py
@@ -142,7 +142,7 @@ def test__interpolated_array_from(grid_2d_7x7):
         pixel_scales=1.0,
     )
 
-    pixelization_grid = aa.Grid2DDelaunay(grid=pixelization_grid_ndarray)
+    pixelization_grid = aa.Mesh2DDelaunay(grid=pixelization_grid_ndarray)
 
     mapper = aa.Mapper(source_grid_slim=grid_2d_7x7, source_mesh_grid=pixelization_grid)
 
@@ -169,7 +169,7 @@ def test__mapped_to_source_from(grid_2d_7x7):
         pixel_scales=1.0,
     )
 
-    pixelization_grid = aa.Grid2DDelaunay(grid=pixelization_grid)
+    pixelization_grid = aa.Mesh2DDelaunay(grid=pixelization_grid)
 
     mapper = aa.Mapper(source_grid_slim=grid_2d_7x7, source_mesh_grid=pixelization_grid)
