+++ test_autoarray/inversion/mappers/test_delaunay.py
@@ -9,7 +9,7 @@ def test__pix_indexes_for_sub_slim_index__matches_util(grid_2d_7x7):
         pixel_scales=1.0,
     )
 
-    pixelization_grid = aa.Grid2DDelaunay(grid=pixelization_grid)
+    pixelization_grid = aa.Mesh2DDelaunay(grid=pixelization_grid)
 
     mapper = aa.Mapper(source_grid_slim=grid_2d_7x7, source_mesh_grid=pixelization_grid)
