+++ test_autoarray/inversion/mesh/test_delaunay.py
@@ -19,7 +19,7 @@ class TestDelaunayMagnification:
             data_grid_slim=sub_grid_2d_7x7
         )
 
-        pixelization_grid = aa.Grid2DDelaunay(
+        pixelization_grid = aa.Mesh2DDelaunay(
             grid=sparse_grid,
             nearest_pixelization_index_for_slim_index=sparse_grid.sparse_index_for_slim_index,
         )
@@ -132,7 +132,7 @@ class TestDelaunayBrightness:
             seed=1,
         )
 
-        pixelization_grid = aa.Grid2DDelaunay(
+        pixelization_grid = aa.Mesh2DDelaunay(
             grid=sparse_grid,
             nearest_pixelization_index_for_slim_index=sparse_grid.sparse_index_for_slim_index,
         )
